requirements.txt
anyio==3.6.2
fastapi==0.95.0
idna==3.4
pydantic==1.10.7
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
alembic==1.7.6
psycopg==3.1.8
fastapi-sqlalchemy==0.2.1
jose==1.0.0
passlib==1.7.4